+++ lib/data/enums/index.dart
@@ -6,4 +6,4 @@ export 'image_size.dart';
 export 'image_style.dart';
 export 'mic_send_mode.dart';
 export 'player_state.dart';
-export 'sender_type.dart';
+export 'role.dart';