+++ lib/data/enums/role.dart
@@ -1,4 +1,4 @@
-enum SenderType {
+enum Role {
   user,
   assistant;
 }