+++ lib/data/models/realtime_events/speech/speech_start.dart
@@ -1,0 +1,11 @@
+import 'package:vit_gpt_dart_api/data/enums/role.dart';
+
+class SpeechStart {
+  final String id;
+  final Role role;
+
+  SpeechStart({
+    required this.id,
+    required this.role,
+  });
+}