+++ lib/repositories/threads_repository.dart
@@ -76,7 +76,7 @@ class ThreadsRepository extends ThreadsModel {
     );
     Map<String, dynamic> data = response.data;
     var newMessage = Message.fromMap(data);
-    message.messageId = newMessage.messageId;
+    message.id = newMessage.id;
     return newMessage;
   }